Rename Form component to FormDisplay component
