+++ client/src/components/Workflow/Editor/Forms/FormSection.vue
@@ -1,9 +1,9 @@
 <template>
-    <Form :id="id" :inputs="inputs" ref="form" @onChange="onChange" />
+    <FormDisplay :id="id" :inputs="inputs" ref="form" @onChange="onChange" />
 </template>
 
 <script>
-import Form from "components/Form/Form";
+import FormDisplay from "components/Form/FormDisplay";
 import makeSection from "./makeSection";
 
 export default {
